Made the importing of files simpler and renamed 'functions' to 'myfunctions' just because it may get in the way of something
